ds-server/requirements.txt
APScheduler==3.8.1
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.11
cos-python-sdk-v5==1.9.15
crcmod==1.7
dicttoxml==1.7.4
Django==4.0.2
django-apscheduler==0.6.0
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
pycodestyle==2.8.0
pycryptodome==3.14.1
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uWSGI==2.0.20
